+++ src/components/Header.jsx
@@ -4,13 +4,13 @@ import todo2 from "../assets/todo2.png";
 import Form from "./Form";
 
 
-const Header = ({task, setTask, handleSubmit, inputRef}) => {
+const Header = () => {
   return (
     <header>
       <div className="title">To Do List</div>
       <div className="logo">
         <img src={todo2} alt="logo" />
-        <Form handleSubmit={handleSubmit} task={task} setTask={setTask} inputRef={inputRef}/>
+        <Form/>
       </div>
       
     </header>
