+++ src/context/MyContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const MyContext = createContext(null)
+
+export default MyContext
