+++ src/index.js
@@ -1,9 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import MyProvider from './context/MyProvider'
 import App from './App';
 
 ReactDOM.render(
-    <App />, document.querySelector('#root')
+    
+    <MyProvider>
+    <App />
+    </MyProvider>
+    , document.querySelector('#root')
 );
